[Chore] 뷰 분리 (#53)
